requirements.txt
# Web框架
fastapi==0.68.0
uvicorn==0.15.0
python-multipart==0.0.5

# 图像处理
Pillow==8.3.1
opencv-python-headless>=4.5.0

# 数据库
sqlalchemy==1.4.22
pymysql==1.0.2

# 云服务
oss2==2.15.0

# AI相关
zhipuai==1.0.7
ultralytics>=8.0.0
# mtcnn>=1.0.0  # 可选依赖，如果使用MTCNN人脸检测

# 配置管理
python-dotenv==0.19.0

# 网络请求
requests>=2.25.0

# 其他工具
pydantic>=1.8.0,<2.0.0
numpy>=1.21.0
tqdm>=4.62.0

# 异步处理
anyio>=3.4.0,<5